Isolate sidebar and mainmenu
